Code simplification
Commented some team/clan actions in python and javascript because
we probably wont need them. Some situations just wont be able to happen
because there will be no buttons for them in the tables on social page.

Automatic clan joining is also not so simple because the clan might already
have a leader which means we must wait for confirmation of join request.
Current solution tries to join immediately and if its not possible, puts player
in the pending tables of specified team and its parent clan.